Update index.html
Both the <h1> and the <ul> elements are set to display: inline-block;, making them sit next to each other horizontally.
The "Contact us" link is placed within a <div> element set to float: right;, which makes it align to the far right of the parent container.
Now, the "ARNOLD" heading and the navigation links are on the same line, and the "Contact us" link is aligned to the far right of the page.